Test register-form reachead from the press or from a journal.
They do use different templates than those used to show the user
profile (which is important) and do have some minor differences (which
does not seem very important).

Using `django.tests.Client` with a URL from the press seem to require
that the press object has the same domain as the fake test
server (which is "testserver") and that the `URL_CONFIG` setting is
set to "domain". Otherwise, the function `get_by_request` defined in
`core.model_utils` and used (I think) by the middleware, will not find
the URL and redirect to the `DEFAULT_HOST` (or, if one adds
`follow=True` to `client.get(...)`, the system will break because or
redirect loops).